backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Database
pymongo
motor
beanie

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pydantic[email]>=2.10
pydantic-settings
slowapi

# AI Integration
pydantic-ai
anthropic

# Email
aiosmtplib

# Testing
pytest
pytest-asyncio
httpx

# Development Tools
ruff
black

# Monitoring
sentry-sdk[fastapi]

# Utilities
python-dotenv
pyyaml